fix(runtime): move hex to crypto
