--- actor-sdk/sdk-core/runtime/runtime-android/src/main/java/im/actor/runtime/HexRuntimeProvider.java
@@ -1,7 +1,0 @@
-package im.actor.runtime;
-
-import im.actor.runtime.generic.GenericHexProvider;
-
-public class HexRuntimeProvider extends GenericHexProvider {
-
-}
