+++ actor-sdk/sdk-core/runtime/runtime-shared/src/main/java/im/actor/runtime/CryptoRuntime.java
@@ -19,4 +19,10 @@ public interface CryptoRuntime {
      */
     @ObjectiveCName("MD5WithData:")
     byte[] MD5(byte[] data);
+
+    byte[] fromHex(String hex);
+
+    byte[] fromHexReverse(String hex);
+
+    String toHex(byte[] raw);
 }
