--- actor-sdk/sdk-core/runtime/runtime-shared/src/main/java/im/actor/runtime/HexRuntime.java
@@ -1,9 +1,0 @@
-package im.actor.runtime;
-
-public interface HexRuntime {
-    byte[] fromHex(String hex);
-
-    byte[] fromHexReverse(String hex);
-
-    String toHex(byte[] raw);
-}
